Added the stub for the workshop, will be used by tinkering to create tools using ideas from other stats, particularly scavenging.  Tools will be used to improve the base stats (armor plating for HP, simple weapons for Attack, and harvesting results)
Also, I think I want to attempt to refactor ALL of the reveal code, it seems to me I continue to duplicate this code and it could be abstracted to follow a naming convention pattern that would allow me to show and hide things as they unlock with a bit of configuration.  Not all panels are completely compatible with this idea, so I want to think on it some more before I go down that road.  Would pretty massively cut down on total code volume though.. so betting it might be worth doing.  Same sorta basic idea as the generic incrementor functions and stuff like that.